app/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
requests